ine/app/requirements.txt
appnope==0.1.2
asgiref==3.3.4
backcall==0.2.0
decorator==5.0.9
Django==3.0
djangorestframework==3.12.4
ipdb==0.13.9
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
Markdown==3.3.4
matplotlib-inline==0.1.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
ptyprocess==0.7.0
Pygments==2.9.0
pytz==2021.1
South==1.0.2
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
wcwidth==0.2.5
gunicorn==20.0.4
psycopg2-binary==2.8.5
